# npu-stack/backend/routers/orchestrator_bridge.py
"""NPU-STACK → Shoggoth Orchestrator dispatch bridge.

Wires the NPU-STACK FastAPI service to the Shoggoth orchestrator for
real compute fabric dispatch. Replaces placeholder stubs in genomic_training.py
with actual HTTP calls to the orchestrator REST API.

Usage (import in genomic_training.py):
    from backend.routers.orchestrator_bridge import OrchestratorBridge

    bridge = OrchestratorBridge("http://localhost:9100")
    decision = await bridge.analyze_workload(code_snippet)
    nodes = await bridge.get_available_nodes("MatrixTensorCore")
"""

from __future__ import annotations

import logging
from dataclasses import dataclass, field
from typing import Optional

import httpx

logger = logging.getLogger(__name__)

# ── Types ──────────────────────────────────────────────────────────────────────


@dataclass
class NodeInfo:
    node_id: str
    tier: str
    vram_gb: int
    ping_ms: float
    accepting_work: bool
    temperature_c: float


@dataclass
class RoutingDecision:
    workload: str
    target_node: str
    reason: str
    suggested_template: str
    template_manifest: str
    confidence: float


@dataclass
class DispatchResult:
    work_id: int
    success: bool
    output_size: int
    elapsed_us: int
    node_id: str
    error: Optional[str] = None


# ── Bridge ─────────────────────────────────────────────────────────────────────


class OrchestratorBridge:
    """Bridges NPU-STACK to the Shoggoth orchestrator REST API.

    All GPU compute dispatch flows through this bridge:
        NPU-STACK (Unsloth training) → OrchestratorBridge → Shoggoth Orchestrator
                                                                      ↓
                                                              Compute Fabric
                                                              (RTX 5090 + MI50 + BC250)
    """

    def __init__(self, base_url: str = "http://localhost:9100", timeout: float = 60.0):
        self.base_url = base_url.rstrip("/")
        self.timeout = timeout
        self._client: Optional[httpx.AsyncClient] = None

    async def __aenter__(self) -> "OrchestratorBridge":
        self._client = httpx.AsyncClient(base_url=self.base_url, timeout=self.timeout)
        return self

    async def __aexit__(self, *args):
        if self._client:
            await self._client.aclose()

    @property
    def client(self) -> httpx.AsyncClient:
        if self._client is None:
            raise RuntimeError("Bridge not opened. Use 'async with OrchestratorBridge()'")
        return self._client

    # ── Health ─────────────────────────────────────────────────────────────────

    async def health(self) -> dict:
        r = await self.client.get("/health")
        r.raise_for_status()
        return r.json()

    # ── Topology ───────────────────────────────────────────────────────────────

    async def get_topology(self) -> dict:
        r = await self.client.get("/topology")
        r.raise_for_status()
        return r.json()

    async def get_available_nodes(self, capability: str) -> list[NodeInfo]:
        """Fetches nodes with a specific capability (e.g., 'MatrixTensorCore')."""
        r = await self.client.get("/fabric/nodes", params={"capability": capability})
        r.raise_for_status()
        data = r.json()
        return [NodeInfo(**n) for n in data["nodes"] if n.get("accepting_work")]

    # ── Analysis ───────────────────────────────────────────────────────────────

    async def analyze_workload(self, source_code: str, project_name: str = "") -> RoutingDecision:
        """Classifies ML/rendering workload and routes to optimal hardware."""
        r = await self.client.post(
            "/analyze",
            json={"source_code": source_code, "project_name": project_name},
        )
        r.raise_for_status()
        data = r.json()
        return RoutingDecision(
            workload=data["workload"],
            target_node=data["target_node"],
            reason=data["reason"],
            suggested_template=data["suggested_template"],
            template_manifest=data["template_manifest"],
            confidence=data["confidence"],
        )

    # ── Dispatch ───────────────────────────────────────────────────────────────

    async def dispatch_training(
        self,
        model_name: str,
        dataset_path: str,
        lora_rank: int = 64,
        epochs: int = 3,
        precision: str = "bf16",
    ) -> dict:
        """Dispatches an Unsloth fine-tuning job to the Shoggoth fabric.

        The orchestrator's agentic parser will:
          1. Detect the training workload type.
          2. Route embedding layers to AMD MI50s.
          3. Route transformer blocks to BC250 APU grid.
          4. Route output heads to RTX 5090.
        """
        source_code = f"""
# Shoggoth dispatch: Unsloth fine-tuning
from unsloth import FastLanguageModel
model = FastLanguageModel.from_pretrained("{model_name}")
# Training config: lora_r={lora_rank}, epochs={epochs}, precision={precision}
"""
        # First, analyze to get routing.
        decision = await self.analyze_workload(source_code, model_name)
        logger.info(
            "Training dispatch: %s → %s (%.0f%% confidence)",
            decision.workload,
            decision.target_node,
            decision.confidence * 100,
        )

        # Launch the heavy-compute template.
        r = await self.client.post(
            "/launch",
            json={"template_name": "heavy-compute", "project_name": model_name},
        )
        r.raise_for_status()
        return r.json()

    async def dispatch_inference(
        self,
        model_name: str,
        input_tokens: list[int],
        max_tokens: int = 256,
    ) -> dict:
        """Dispatches an inference request to the Shoggoth fabric."""
        r = await self.client.post(
            "/launch",
            json={
                "template_name": "heavy-compute",
                "project_name": f"{model_name}-inference",
            },
        )
        r.raise_for_status()
        return r.json()

    # ── Cloud Provisioning ─────────────────────────────────────────────────────

    async def provision_cloud_nodes(
        self, capability: str = "MatrixTensorCore", count: int = 2
    ) -> dict:
        """Requests cloud GPU provisioning if local nodes are saturated."""
        r = await self.client.post(
            "/fabric/register",
            json={
                "node_id": f"cloud-provision-request",
                "tier": "CloudScale",
                "capabilities": [capability],
                "available_vram_gb": 80,
                "network_ping_ms": 8.5,
                "accepting_work": True,
                "temperature_c": 45.0,
            },
        )
        r.raise_for_status()
        return r.json()